ipc: Fix dynamic irq resource request for specific host
The board config for Linux uses the specific host instead
of HOST_ID_ALL. Also, the board config has updated the
range values for C7x, so that the offset needs to be
updated to fit within the new range.

This patch adds code to first try with the specific host,
and only if that fails, try with HOST_ID_ALL. In this
way it can support both board configs.

It also updates the interrupt selection offset for C7x to
be 2 instead of 51, to fit within the allowed range with
the Linux board config and updated RTOS board config.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>