SBL: OSPI: Fix instability for CPU reads from OSPI flash
Enable Dual Transfer Rate (DTR) mode of OSPI interface for
the SBL to safely copy the SYSFW image out of OSPI flash.
There was instability in reading the SYSFW image without
this setting, especially after the board had been on for
some time or after multiple power cycles.

This is a workaround for bug PRSDK-8301, until such time
as SDR mode of the OSPI interface can be made stable again
for all image transfers.

NOTE: this workaround causes a boot time degradation of
about 2.6 ms for the SBL CUST build (OSPI boot) when using
OSPI in DDR mode, instead of in SDR mode, since DMA transfers
cannot be used early in the SBL to transfer the SYSFW binary.

Signed-off-by: Jonathan Bergsagel <jbergsagel@ti.com>