SBL: add flag to enable HLOS to control OSPI flash
Adds build flag, SBL_HLOS_OWNS_FLASH, to allow the SBL
'CUST' build to bypass setting up the OSPI flash in
XIP mode.  This allows the subsequent Boot App (used to
boot HLOS on Cortex-A cores) to boot the required cores
and then reconfigure the OSPI flash in SPI mode in order
for the HLOS to take control of the flash.

Signed-off-by: Jonathan Bergsagel <jbergsagel@ti.com>