PDK-5733: SBL: Updates SBL_IsSysfwEnc to use devicetype
In older sysfw releases, there was no way to tell whether the SYSFW was
encrypted or not, without parsing the sysfw to check if it was dual
signed.

After SYSFW Rel 2019.11, it is possible to read the device type from SoC
registers - this enables a faster way to determine if the sysfw is
encrypted or clear text.

This patch updates the SBL_IsSysfwEnc to return that the sysfw is clear
text, only if the device type is GP.

Signed-off-by: Madan Srinivas <madans@ti.com>