sciclient: Fix IR output validation against board config
Fix the interrupt routing validation so that
only routes with IR outputs that validate
against the RM board configuration are used.

The logic was in place for this before but
an error in the code allowed routes to still be
programmed even if the validation against the
board configuration failed.

Fixes: PDK-5768

Signed-off-by: Justin Sobota <jsobota@ti.com>