PDK-5733: SBL: Removes roundabout way to determine device type
Earlier, the secmgr registers that store the device type was accessible
only to DMSC M3. So the SBL determined the device type by reading teh
sysfw and checking if it was countersigned and encrypted.

From SYSFW 2019.11 the secmgr registers are made avaiable to the
non-secure software and SBL_IsSysfwEnc has been updated to return the
device type from the secmgr registers.

Hence this workaround of parsing the sysfw certs to determine the device
type must be removed.

Signed-off-by: Madan Srinivas madans@ti.com