[PDK-5988] UDMA : Fix Bug introduced by SA issue resolution
-Fix for a minor bug introduced by Static Analysis issue resolution in commit:97fecfcd

-In case of LCDMA, since Teardown event is not supported, the eventRegister &  eventUnregister will be bypassed.
As a result eventInitDone wont be set to UDMA_INIT_DONE.
While bypassing eventUnregister, as part of SA fix, the check for eventInitDone was moved  alongwith Handle NULL_PTR check.
It was on top of the if condition to bypass the Teardown event. Due to this, Teardown event returned EFAIL in AM64x.

-The check for eventInitDone is moved under, bypass check.

Signed-off-by: Don Dominic <a0486429@ti.com>