PDK-6454: SBL: Adds support for non-GP device types
There are variants on non-GP device types that dont require app
authentication as the SMPK/BMPK keys are not present.

This patch renames the SBL_IsSysfwEnc function to SBL_IsAuthReq to
better reflect this functionality and and enables it to detect
all device types that dont need authentication.

Authentication is needed, and indeed, only possible, on the EMU/HS
device type, SE subtype. These are basically basically EMU ahd HS
devices with a valid SMPK or BMPK key for authenticating the app
against.

Signed-off-by: Madan Srinivas <madans@ti.com>