PDK-6624: WORKAROUND for SBL OSPI Boot instability on J721E-HS
WORKAROUND: On J721E-HS, the signed application is copied into DDR
where it is verified to be signed correctly. However, with OSPI
boot on the J721E-HS EVM, the verification is not proceeding in DDR.
If we change SBL_SCRATCH_MEM_START to use MSCM RAM, instead, then
it consistently boots fine.

The DDR stability should be fixed in CSL.