PDK-6738: IPC: echo_testb: Copy vecs for ipc_echo_testb sysbios example
The latest sysbios expects the vecs to be present at address 0x0.
When the ipc_echo_testb starts from BTCM from SPL branch-only mode
for mcu1_0, the vecs at 0x0 are not matching the sysbios vecs.

This patch adds a function to the sysbios reset functions to
copy the vecs to 0x0 during bios startup for the btcm image for
mcu1_0. With this, the vecs are as expected for sysbios and the image
is able to run sucessfully and finish the echo test, even when
run in branch-only mode.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>