PDK-5593: IPCLLD: Support OSAL Abstraction for Baremetal
Previously OSAL params for baremetal had to be completely
provided by the application. A proper osal abstraction supported
directly from IPC LLD contributes to ease of use for the
application. The osal params can still be overridden by
the user-supplied params, but in case nothing is provided,
defaults are selected for baremetal case.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>