PRSDK-6981: MB LLD: AM64x: Updates to support usage from IPC LLD
To support usage from IPC LLD, checks for certain osal
functions used by Mailbox LLD are moved to be checked
only if they are to be used by Mailbox LLD. For example,
certain osal functions were used only by TPR12, or only
for certain write/read modes. The checks for valid
function is then moved to be checked only if they will
be used.

Additionally, the missing Mailbox_deinit API is added,
and a minor documentation fix for the osal registerIntr
function.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>