PDK-7142: IPC: Update J7VCL Shared Memory Address
For J7VCL, move the location of the shared memory buffer
closer to the carveouts used for loading the firmwares.
This reduces the gap between the regions used for firmware
loading and shared memory for the R5F cores.

New location is at 0xA4000000 for J7VCL.
J721E remains at 0xAA000000.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>