PDK-7061: Mailbox LLD: Single Lib for RTOS and Baremetal
Update the mailbox lld to remove the separate mailbox and mailbox_baremetal
libraries. Instead, have a single library, that can be linked with either
tirtos osal or nonos osal by application.
In order to do this, the mailbox library uses a new OSAL API that is used
for am64x R5F direct VIM interrupt registration (new API is used in certain
readMode by am64x mailbox implementation only).
Additionally, add a new open param "enableVIMDirectInterrupt", which can
be used to enable/disable this feature. By default, it is set to false,
but can be enabled by the application.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>