PDK-5611: SBL: Switch OSPI flash to SPI mode at the end
Adds support for re-opening the OSPI flash in SPI mode
and leaving it in that mode when the SBL exits.

This supports U-boot or HLOS boot on the Cortex-A cores
of the SoC. It allows the SPL/U-boot to take over the
flash, reset it, and reconfigure it for Octal-mode again.

Signed-off-by: Jonathan Bergsagel <jbergsagel@ti.com>