PDK-7536: ipc_examples: Flush the Linux trace buffer upon Exception
The c6x trace buffer in the ipc examples is in cached memory. Since it
is in cached memory, in order to see the contents of the SysBIOS exception
traces post-crash, the trace buffer needs to be written back to DDR so
that it can be visible from Linux running on A72. Being able to view
the exception traces from Linux command-line is desirable as access to
CCS to view the trace buffer is not always available.

This patch plugs an Exception hook that flushes the contents of the trace
buffer when an exception occurs.
Currently there isn't an IPC example that simulates a crash on C6x, but
adding this gives an example of how to do this when developing applications
with IPC that communicate with Linux on A72.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Keerthy <j-keerthy@ti.com>