PDK-7060: IPC: Update the SysBios timer used for J7200 examples
Power for the main domain timer 0 to 7 is controlled by GP_CORE_CTL
and others are controlled by LPSC_PER_SPARE_0.

By default the GP_CORE_CTL is powered up SBL/GEL's
So, configure to use the timers 6 & 7 for main R5F's.

This aligns with the defaults selected in PDK default config for
J7200.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>