PDK-7625: IPC: Fix issue with delayed announce message
The IPC LLD should not clear the messages from the mailbox
when starting up, as another core may have come up before
and already sent a message to the mailbox. Instead, move
the enabling of the interrupt to later once the rpmsg is
initialized and is able to handle the pending message via
the callback. In this way, if an announce message is
already sent to the core, it will be processed as soon
as the core is ready to accept messages and not delayed
until the next message is received.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>