PDK-7687: IPC: Fix negative test timeout test
The timeout test was incorrectly returning fail when it should have
been returning pass. Fix it to return the proper status.

Also, init the timer interrupts for the negative test for
C66x and C7x because it is needed for testing the timeout functionality,
and also the test uses Task_sleep.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>