Fix the SBL R5F-0 cluster reset sequence
SBL has been illegally resetting the R5F-0 cluster.

Core 1 must never be in a higher functional state than core 0, and it is a bad idea to power-down (and power back on) the cores once TCMs have been loaded (as the data will not necessarily be valid when the RAM comes back on-line).