PDK-7757: PSDKQA-198: QNX: Fix the IPC error
New warning is enabled in QNX build that throws an
error when strncpy is used with length less than the
size of the length of the source string. Although in
this case, the next line was already taking care to
make sure the string is NULL-terminated.
Fix the compilation error by using length that is
the same size as the source name entry since the source
and destination name entries are the same size.

Signed-off-by: Praveen Rao <prao@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>