PDK-8188: Mailbox: Save per-run results to unique address and print results
Update the perf test results to print the average round-trip based on
the per-run latencies instead of the average over all the runs.
This gives a more accurate value. Additionally, save the per-run
information for each core combination so that it can be copied off
the target, if desired, and further analyzed.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>