PDK-8120: IPC: Fix j7200 ipc_perf_test vring settings
The build for j7200 for the ipc_perf_test was incorrectly
using the j721e vring settings, so the correct region
was not being set as un-cached.

This patch fixes the makefile to use the correct config
file. Also, the files are updated for mcu1_0 to be able
to be loaded from CCS or sbl by adding the copyVecsToAtcm.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>