PDK-8155: IPC: test: Add sleep while waiting for Linux
Between checking for the remote (Linux) to be ready via the
call to Ipc_isRemoteReady, the test should sleep the task
to give a chance for other tasks to run rather than
busy waiting for Linux.

To enable Task_sleep to be functional for all cores, also
enable the timer interrupts for C7x and C66x, rather than only
for the IPC negative test.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>