PDK-8180: Mailbox: Add new API for Enable/Disable Interrupts
In some use cases it is required for the application to be able
to delay enabling of the mailbox interrupts so that further
setup can be done before it starts handling the interrupts.

One such example is when IPC LLD uses Mailbox LLD, which needs
to perform some setup after opening the Mailbox handle.

To support this use case, the following APIs are added:

Mailbox_enableInterrupts
Mailbox_disableInterrupts

Also, the following new open-time param is added:

enableInterrupts

The value of enableInterrupts defaults to true so backward
compatibility is maintined.

Currently only AM64X is supporting this feature and new
APIs.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>