PDK-8215: SBL: Update OSPI/Hyperflash layout of images
Bootloader image sizes are larger now, with the added code
for sciclient and rm_pm_hal, due to required HSM rearch.

SBL image size was going beyond its original allotted space
in the flash, so we have to update the flash locations to
the following now:

Location   Image
--------   ------------------------
00_0000    SBL (tiboot3.bin)
08_0000    SYSFW (sysfw.bin)
10_0000    APP (*.appimage)
1C_0000    (optional) XIP boot test

This also allows us to align image locations with multiples
of 256KB sections in the flash for better flash arrangement.

Signed-off-by: Jonathan Bergsagel <jbergsagel@ti.com>