sciserver: forward non-rm/pm message to TIFS
If the message type is not RM/PM, then it needs to be forwarded to TIFS.
The full message size is sent/received from TIFS because we do not parse
the message type to identify the actual message size.

This should resolve the issue where Processor Control messages are
dropped.

Fixes: PDK-8299 #integrate-and-build

Signed-off-by: Sheng Zhao <shengzhao@ti.com>