Fix for the state capture for both output and input IR output for ROM mapping
The Sciclient RM IRQ has a logic which checks if the input and output of an IR is mapped in the ROM table or not. The way the logic works is that at the Sciclient level the mapping is checked but not cleared. However, a flag which says the actual entries are cleared is set to true indicating it is cleared. This is fine if only one of the input or output is mapped.

if both are mapped the logic in the Sciclient_rm_irq.c file fails to recognise the rom mapping for the output since a single flag which indicates the input is used is already set. This causes a confusion and the RM IRQ eventually fails.

This patch will track the output and input separately.

Fixes: PDK-9005

Signed-off-by: Piyali Goswami <piyali_g@ti.com>