Fix for UDMA tests failing after fix for PDK-9013
Fixes in Sciclient.c where the respHdr is not populated in the respPayload.
The Sciserver logic looks for the payload to have the flags set correctly and because the flags were never copied in the case when the message was forwarded to the TIFS, the response is falsely reported to the calling function.

Signed-off-by: Piyali Goswami <piyali_g@ti.com>