ETHFW-1765 - timesync: v2: Fix incorrect packet state check in TimeSync_reset()
EnetAppUtils_validatePacketState() was causing an assertion indicating
that a packet in rxFreeQ wasn't in the expected state (FREEQ state).

EnetAppUtils_validatePacketState() checks that packets in a queue are in
FREEQ state and moves them to DRIVER state. So it's a logical error
to call it multiple times on the same queue while packets are being
added to the queue.

Essentially, the packets which were already in the queue and have
already being validated via EnetAppUtils_validatePacketState() will
always fail a second validation as they have already moved to DRIVER
state.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>