Disabling PTP Tx timestamp check feature
A safety feature was added for V1 ICSS to make sure that Tx timestamp has indeed changed and this was carried over to V2 ICSS like AM5. In this we have a state machine to
compare current and previous timestamps and make sure that timestamps are indeed different. This check is not required for AM5 because we push more number of bytes (>22) before PTP timestamp is taken. This helps save cycles.