board: am65xx: Use PDK-relative path for pinmux.h
If an application wants to set their own pinmux, it has to include board
library's internal am65xx_idk_pinmux.h header file as the header files
in its component include/ directory are not enough.

The internal header files included in am65xx_idk_pinmux.h would need to
include the pah relative to PDK, otherwise applications will not be able
to find them.  That's the case of pinmux.h which is included, but path
is not relative to PDK base dir.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>