PDK-9712: Fix for multiple aborts in Sciclient Firewall test app
The Sciclient firewall test app checks for reads and writes when trying to check for the firewall based data aborts.
The SYSBIOS behavior is that it generated data aborts only for reads and the Free RTOS behavior is that is generates for reads and writes.
The test is simplified to suit both settings to only try reads and then check the number of aborts.

Signed-off-by: Piyali Goswami <piyali_g@ti.com>