PDK-10280: Fix for C7x interrupt based Sciclient Operation
The C7x CLEC will continue to forward the Secure Proxy interrupt even when the ESR is cleared.
The C7x HWI interrupt handler will re-enable ESR causing the interrupt to keep happening and thus the code to be stuck.
This fix will disable forwarding the interrupt from the CLEC and re-enable forwarding when the interrupt is serviced.

Signed-off-by: Piyali Goswami <piyali_g@ti.com>