IPC: virtio: Make Ipc_isRemoteVirtioCreated public
Added function prototype for Ipc_isRemoteVirtioCreated in ipc_virtio.h
to provide a way to check for existing virtio before calling
IPC_lateVirtioCreate. This is required since IPC_lateVirtioCreate
returns failure (which could be harmless) if the specified remote virtio
already exists. Ideally, IPC_lateVirtioCreate should return an
appropriate error code for this case.

Please refer to ETHFW-1575 for more details on the need to expose this
function.

Signed-off-by: Nitin Sakhuja <n-sakhuja@ti.com>