ETHFW-1922 - board: j721e: Don't set RMII PHY reset/interrupt pins
gJ721E_MainPinmuxDataGesiCpsw9gQsgmii structure is not meant for RMII
usage, so it doesn't enable RMII8 data pins but it was still enabling
RMII PHY's reset and interrupt lines which caused conflicts with McASP
pins.

It's worth noting that EthFw is the user of this pinmux structure and
RMII is not part of J721E's default port configuration.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>