PDK-12485 - osal: safertos: ClockP: Initialize xTaskToNotify to avoid conflict with pxCallbackFunction
SafeRTOS timer can be provided with either a task or callback function,
but not both which is treated as an invalid parameter error.

The OSAL implementation for SafeRTOS uses pxCallbackFunction, hence it
must not use task by explicitly clearing xTaskToNotify timer parameter.
xTaskToNotify was not explicitly set which caused ClockP_create().

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>