J721S2: Update Indirect trigger address location.
The Indirect AHB Trigger Address for j721s2 is currently programmed as 0x50000000 + 0x0U which is cachable, so all the memcpy writes may not landing up in actual memory. Location of SRAM will start from this trigger address. Making this address to point to a non-cachable location fixes the problem.

Fixes: PDK-12255

Signed-off-by: Lohith Kumar <l-kumar@ti.com>