Fix: XSPI: Incorrect programming of baudrate divider
Acc. to OSPI spec the baudrate divider minimum needs to 4 for SDR
and 8 for DDR modes. But currently these values are getting hard-coded
to 6 if frequency < 200MHz which is not in sync with spec.

- Remove hardcodings in driver for baudRateDivider.
- Update Testcases to validate reads at baudRateDiv 8U.

For XSPI flash the read data capture delay needs to be half the
value of baudRateDiv.

- Update nor_xspi.c to program rdDataCapDelay value to baudRateDiv/2.

After fixing baudRateDiv the first byte was getting read incorrectly
while performing INDAC reads.

- Reduce the number of read dummy cycles by 1 for INDAC reads.

Fixes: PDK-13178

Signed-off-by: Lohith Kumar <l-kumar@ti.com>