lwip: port: Restructure SoC and OS configs
Previously, lwIP configuration parameters were structured per OS.
So a given platform had one lwipopts.h header file for each supported
OS with the majority of the configuration parameters set to the same
values.

The right structure is to split the configuration parameters in
OS-dependent and SoC-dependent.  For a given SoC, a single SoC header is
maintained, but different OS headers as needed.

The file organization in the new structure is:
- lwip-port/<OS>/include/lwipopts_os.h
- lwip-port/config/lwipopts_common.h
- lwip-port/config/<SOC>/lwipopts.h

with lwiptops.h including lwipopts_common.h and lwipopts_os.h.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>