[PDK-13207][Sciclient]: Add testcase for PVU to Main R5 core interrupt routing
Testcase is added to program the interrupt route between PVU and R5 core
of main domain. The testcase also generates an interrupt from PVU and
checks if the corresponding ISR is executed by the R5 core.

Testcase also deletes the set interrupt route between PVU and Main R5
and checks if the R5 core still executes the registered ISR.

Signed-off-by: Thejasvi Konduru <t-konduru@ti.com>