[PDK-13154][Sciclient]: Fix Main UART to MCU R5 core interrupt routing testcase
In order to execute the UART testcase, sciclient_unit_testapp is run on
mcu1_0 core, the testcases previous to the UART testcase uses MCU uart
instance and the UART testcase uses MAIN uart instance.

For J7200, J721E the value of MCU uart instance and Main uart instance used
by the board is the same. Both of these instances use the same structure to
store their respective UART Hardware attributes. This structure attributes are
not set to the default values before being used by the MAIN uart instance.
This resulted in the failure for J7200, J721E.

The issue is fixed by a using a different structure to store the values
corresponding to our main uart instance instead of modifying it in the
structure used by the mcu uart instance.

Signed-off-by: Thejasvi Konduru <t-konduru@ti.com>