[Compiler][PDK-13499]: Remove flto flag in rules_ti_cgt_arm.mk
With compiler migration to 3.2.0.LTS,IPC_S_FUNC_RPMSG_SAMPLE_CLIENT
test fails in linux DM for j7200 device.

The root cause of this problem is the link-time optimization during run-time.

To resolve this issue, remove the -flto flag in the rules_ti_cgt_arm makefile.
This flag should be removed for all Jacinto devices to prevent any further problems.

Signed-off-by: Chandru Dhavamani <chandru@ti.com>