[PDK-13500][Sciclient]: Resolve incorrect disable of write protection in Sciclient_serviceSecureProxy API
gSciclient_writeInProgress is a binary semaphore which is used to control the access
of critical section in Sciclient_serviceSecureProxy(). Currently even after this
semaphore variable is disabled, the global variables such as pSciclient_secProxyCfg
in Sciclient_readThread32(), gSciclientHandle are accessed by the threads.

As the threads share the global variables, this makes the code thread unsafe.

Hence shifted the disabling of gSciclient_writeInProgress variable to a correct
position in the Sciclient_serviceSecureProxy().

Signed-off-by: Thejasvi Konduru <t-konduru@ti.com>